[DEV] Refactor/code structure (#49)
* feat(service): Add destination detail service layer

* feat(error): Add error layout for destination detail

* feat(props): Separate props for destination detail

* feat(loading): Add loading page for destination detail

* refactor(page): Separate logic, state and component

* feat: Add destination detail component

* feat(hook): Create custom hooks for destination detail
