+++ app/(route)/destinations/loading.tsx
@@ -1,0 +1,9 @@
+const Loading = () => {
+  return (
+    <div className='flex h-screen items-center justify-center'>
+      <div className='h-10 w-10 animate-spin rounded-full border-4 border-t-4 border-gray-200 border-t-blue-500'></div>
+    </div>
+  )
+}
+
+export default Loading
